+++ app-tauri/assets/index-B-HNoDsN.js
@@ -1,1 +1,0 @@
-import"./core-B8a0kdWx.js";import{listen as e}from"./event-e4m6lkuv.js";async function i(n){return await e("deep-link://new-url",r=>{n(r.payload)})}export{i as onOpenUrl};