+++ app-tauri/assets/index-CaDNzvyl.js
@@ -1,0 +1,1 @@
+import"./core-uYr82bVc.js";import{listen as e}from"./event-D7m5suv4.js";async function i(n){return await e("deep-link://new-url",r=>{n(r.payload)})}export{i as onOpenUrl};