+++ app-tauri/assets/index-g2WdMWDK.js
@@ -1,1 +1,0 @@
-import{i as o}from"./core-B8a0kdWx.js";async function p(i,n){await o("plugin:shell|open",{path:i,with:n})}export{p as open};